+++ engine/flink/util/src/main/scala/pl/touk/nussknacker/engine/flink/util/listener/NodeCountMetricListener.scala
@@ -2,8 +2,7 @@ package pl.touk.nussknacker.engine.flink.util.listener
 
 import cats.data.NonEmptyList
 import pl.touk.nussknacker.engine.api._
-import pl.touk.nussknacker.engine.flink.util.metrics.WithMetrics
-import pl.touk.nussknacker.engine.util.metrics.{Counter, MetricIdentifier}
+import pl.touk.nussknacker.engine.util.metrics.{Counter, MetricIdentifier, WithMetrics}
 
 class NodeCountMetricListener extends EmptyProcessListener with WithMetrics {
