+++ engine/flink/util/src/main/scala/pl/touk/nussknacker/engine/flink/util/metrics/FlinkMetricsProviderForScenario.scala
@@ -81,13 +81,4 @@ class FlinkMetricsProviderForScenario(runtimeContext: RuntimeContext) extends Me
 
 }
 
-trait WithMetrics extends Lifecycle {
 
-  @transient protected var metricsProvider : MetricsProviderForScenario = _
-
-  override def open(context: EngineRuntimeContext): Unit = {
-    super.open(context)
-    this.metricsProvider = context.metricsProvider
-  }
-
-}
