+++ utils/util/src/main/scala/pl/touk/nussknacker/engine/util/exception/ExceptionRateMeter.scala
@@ -2,22 +2,23 @@ package pl.touk.nussknacker.engine.util.exception
 
 import cats.data.NonEmptyList
 import pl.touk.nussknacker.engine.api.exception.EspExceptionInfo
-import pl.touk.nussknacker.engine.util.metrics.RateMeter
+import pl.touk.nussknacker.engine.util.metrics.{InstantRateMeterWithCount, MetricsProviderForScenario, RateMeter}
 
-trait ExceptionRateMeter {
+class ExceptionRateMeter(metricsProvider: MetricsProviderForScenario) {
 
   private lazy val allErrorsMeter: RateMeter = instantRateMeter(Map(), NonEmptyList.of("error", "instantRate"))
 
   private val nodeErrorsMeterMap = collection.concurrent.TrieMap[String, RateMeter]()
 
-  def instantRateMeter(tags: Map[String, String], name: NonEmptyList[String]): RateMeter
-
-  protected def markException(exceptionInfo: EspExceptionInfo[_ <: Throwable]): Unit = {
+  def markException(exceptionInfo: EspExceptionInfo[_ <: Throwable]): Unit = {
     allErrorsMeter.mark()
     getMeterForNode(exceptionInfo.nodeId.getOrElse("unknown")).mark()
   }
 
   private def getMeterForNode(nodeId: String): RateMeter =
     nodeErrorsMeterMap.getOrElseUpdate(nodeId, instantRateMeter(Map("nodeId" -> nodeId), NonEmptyList.of("error", "instantRateByNode")))
+
+  private def instantRateMeter(tags: Map[String, String], name: NonEmptyList[String]): RateMeter =
+    InstantRateMeterWithCount.register(tags, name.toList, metricsProvider)
 
 }
